웹뷰 플로우 wip (#46)
* bridge 이동하는 flow 생성

* main에서 navigate할 수 있는 환경 세팅

* project 생성과 isMe에 따라 나의 태스크 분기 할 수 있는 환경 생성

* projectid int로 변경

+++ presentation/src/main/java/com/yapp/timitimi/presentation/ui/main/MainViewModel.kt
@@ -17,10 +17,10 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(
     private val mainMiddleware: MainMiddleware,
     private val userPreference: UserPreference
-): BaseViewModel<MainIntent, MainState, MainSingleEvent>() {
+) : BaseViewModel<MainIntent, MainState, MainSingleEvent>() {
     init {
         start()
-        dispatch(MainIntent.Init(userPreference.lastViewedProjectId))
+        dispatch(MainIntent.Init(userPreference.lastViewedProjectId.toInt()))
     }
 
     override fun registerMiddleware(): List<BaseMiddleware<MainIntent, MainSingleEvent>> {
